+++ NWNXLib/API/FunctionsLinux.hpp
@@ -8,7 +8,7 @@ NWNXLIB_FUNCTION(_ZN10CNWSPlayerD0Ev)
 NWNXLIB_FUNCTION(_ZN10CNWSPlayerD1Ev)
 NWNXLIB_FUNCTION(_ZN10CNWSModule15LoadModuleStartE10CExoStringiiRKN6NWSync13AdvertisementE)
 
-// Hooking functions that override a virtual functions doesn't work with the new &Class::Function way...
+// Hooking functions that override a virtual function doesn't work with the new &Class::Function way...
 // .. so we need these!
 NWNXLIB_FUNCTION(_ZN11CNWSTrigger12EventHandlerEjjPvjj)
 NWNXLIB_FUNCTION(_ZN12CNWSCreature12EventHandlerEjjPvjj)
@@ -36,5 +36,7 @@ NWNXLIB_FUNCTION(_ZN9CNWSStore8AIUpdateEv)
 
 NWNXLIB_FUNCTION(_ZN21CNWSEffectListHandler15OnEffectAppliedEP10CNWSObjectP11CGameEffecti)
 NWNXLIB_FUNCTION(_ZN21CNWSEffectListHandler15OnEffectRemovedEP10CNWSObjectP11CGameEffect)
+
+NWNXLIB_FUNCTION(_ZN16CNWSAmbientSound15PlayBattleMusicEi)
 
 // ***